+++ packages/app-store/src/index.ts
@@ -1,3 +1,1 @@
-// Export everything from appStore
 export * from './appStore'
-export * from './types'
